Drone corectely moving to waypoints
Drone now get's a new waypoint once it's reached the one it's currentely
at.  Going to build a drone and turret and player factory before moving
further with drone so game manager can keep up with who's in the scene.
so the turret and drone may more appropriately get targets.